Add files via upload
I created a Student Grade Calculator to simplify evaluating academic performance. It inputs marks for subject and it calculates the total marks and average percentage, then assigns a corresponding grade. The results—total marks, average percentage, and grade—are displayed to the user, making the grading process quick and accurate.